Initial implemetation of evaluation of let and lambda expressions
Add a local variable context to the interpreter context. Let expressions
just add to this, shadowing is handled by replacing entries in the
hashmap.

Lambda function evaluation does not implement lexical scoping
because I don't want to deal with captured values yet.

Currently in variable name resolution first the local conctext is
checked, then cells, and if neither of these contains the name then the
name is assumed to be a buiiltin function. I want to swap this around so
that cell names are last in the order because this means I don't have to
deal with the case where a cell is renamed to a builtin function and
would trigger a change in resolution of all cells referencing that
builtin function.