Rework let expressions and sequencing
Removed sequencing operations and let is no longer an expression by it
self, rather it introduces a new scope with some bindings and then is
followed by an expression that is evaluated in that scope. The scoping
rules for sequencing with let expressions were a bit unintuitive at
times so I opted for a more classic approach. A new scope is created
with a let expression.

Each let expression can bind multiple names but this is treated as a
series of consecutive let expressions in that expressions can only refer
to values named before them.